+++ spec/active_record_proxy_adapters/postgresql_proxy_spec.rb
@@ -25,11 +25,13 @@ RSpec.describe ActiveRecordProxyAdapters::PostgreSQLProxy do # rubocop:disable R
     SQL
   end
 
+  # rubocop:disable RSpec/MultipleMemoizedHelpers
   shared_examples_for "a_proxied_method" do |method_name|
     subject(:run_test) { proxy.public_send(method_name, sql) }
 
     let(:proxy) { described_class.new(primary_adapter) }
     let(:read_only_error_class) { ActiveRecord::ReadOnlyError }
+    let(:model_class) { TestHelper::PostgreSQLRecord }
 
     context "when query is a select statement" do
       let(:sql) { "SELECT * from users" }
@@ -74,7 +76,7 @@ RSpec.describe ActiveRecordProxyAdapters::PostgreSQLProxy do # rubocop:disable R
         it "reroutes query to the primary" do
           allow(primary_adapter).to receive(:"#{method_name}_unproxied").and_call_original
 
-          ActiveRecord::Base.connected_to(role: TestHelper.writing_role) { run_test }
+          model_class.connected_to(role: TestHelper.writing_role) { run_test }
 
           expect(primary_adapter).to have_received(:"#{method_name}_unproxied").with(sql, any_args).once
         end
@@ -82,7 +84,7 @@ RSpec.describe ActiveRecordProxyAdapters::PostgreSQLProxy do # rubocop:disable R
         it "does not checkout a connection from the replica pool" do
           allow(replica_pool).to receive(:checkout).and_call_original
 
-          ActiveRecord::Base.connected_to(role: TestHelper.writing_role) { run_test }
+          model_class.connected_to(role: TestHelper.writing_role) { run_test }
 
           expect(replica_pool).not_to have_received(:checkout)
         end
@@ -109,7 +111,7 @@ RSpec.describe ActiveRecordProxyAdapters::PostgreSQLProxy do # rubocop:disable R
       context "when sticking to replica" do
         it "raises database error" do
           expect do
-            ActiveRecord::Base.connected_to(role: TestHelper.reading_role) { run_test }
+            model_class.connected_to(role: TestHelper.reading_role) { run_test }
           end.to raise_error(read_only_error_class)
         end
       end
@@ -153,6 +155,7 @@ RSpec.describe ActiveRecordProxyAdapters::PostgreSQLProxy do # rubocop:disable R
       end
     end
   end
+  # rubocop:enable RSpec/MultipleMemoizedHelpers
 
   describe "#execute" do
     it_behaves_like "a_proxied_method", :execute do